+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add loan-action filter options In Use and Held. Refs UICIRCLOG-181.
 * Migrate jest to use `@folio/jest-config-stripes`. Refs UICIRCLOG-182.
 * Add `stripes-core.settings.read` permission to app permissions. Refs UICIRCLOG-190.
+* Reset offset when applying new search filters. Refs UICIRCLOG-185.
 
 ## 6.0.2 (https://github.com/folio-org/ui-circulation-log/tree/v6.0.2) (2025-11-04)
 [Full Changelog](https://github.com/folio-org/ui-circulation-log/compare/v6.0.1...v6.0.2)